chore(metabase): rename MetabaseAccessSyncer to MetabaseAccessRemover
